Simple website for displaying games
I wanted to create a simple website that displays some games and their description. I may possibly expand upon this and use JS to create my tiles by fetching a game object that contains a bunch of game names and then gets that data from google to display a collection of random tiles.